test: refactor test package structure to correspond to source code package structure
