+++ src/test/java/edu/ntnu/idatt2003/gruppe50/domain/market/StockTest.java
@@ -1,6 +1,5 @@
-package edu.ntnu.idatt2003.gruppe50.model;
+package edu.ntnu.idatt2003.gruppe50.domain.market;
 
-import edu.ntnu.idatt2003.gruppe50.domain.market.Stock;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
