+++ src/test/java/edu/ntnu/idatt2003/gruppe50/domain/trade/TransactionFactoryTest.java
@@ -1,11 +1,7 @@
-package edu.ntnu.idatt2003.gruppe50.transaction;
+package edu.ntnu.idatt2003.gruppe50.domain.trade;
 
 import edu.ntnu.idatt2003.gruppe50.domain.portfolio.Share;
 import edu.ntnu.idatt2003.gruppe50.domain.market.Stock;
-import edu.ntnu.idatt2003.gruppe50.domain.trade.Purchase;
-import edu.ntnu.idatt2003.gruppe50.domain.trade.Sale;
-import edu.ntnu.idatt2003.gruppe50.domain.trade.Transaction;
-import edu.ntnu.idatt2003.gruppe50.domain.trade.TransactionFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
