+++ src/test/java/edu/ntnu/idatt2003/gruppe50/shared/observer/ObservableTest.java
@@ -1,9 +1,8 @@
-package edu.ntnu.idatt2003.gruppe50.observer;
+package edu.ntnu.idatt2003.gruppe50.shared.observer;
 
 import edu.ntnu.idatt2003.gruppe50.domain.market.Exchange;
 import edu.ntnu.idatt2003.gruppe50.domain.market.Stock;
 import edu.ntnu.idatt2003.gruppe50.domain.trade.TransactionFactory;
-import edu.ntnu.idatt2003.gruppe50.shared.observer.Observer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
